requirements.txt
httpx==0.28.1
openai==2.32.0
pydantic==2.13.3
python-dotenv==1.2.2

PyYAML==6.0.3
